project/requirements/base.txt
# Bleeding edge Django
django==1.8.5
Jinja2==2.8.1
django-jinja==2.2.2
markupsafe==0.23

# Configuration
django-environ==0.4.1
django-secure==1.0.1
whitenoise==3.2.2

# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1
django-floppyforms==1.7.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
git+git://github.com/mozilla/unicode-slugify#egg=unicode-slugify==0.1.3-HEAD
django-autoslug==1.9.3

# Time zones support
pytz==2016.10

# Your custom requirements go here
django-bootstrap3==7.1.0
django-reversion==1.9.3
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==0.15.3
djangorestframework-filters==0.9.1
django-markdown==0.8.4
dateutil==2.4.2
django-settings-export==1.2.1
holviapi==0.2.20170129
python-dateutil==2.6.0
django-phonenumber-field==1.1.0
